requirements.txt
Django>=5.0,<6.0
graphene-django>=3.0
django-graphql-jwt>=0.4.0
django-filter>=24.0
psycopg2>=2.9
celery>=5.3
redis>=5.0
Pillow>=10.0
gunicorn>=21.0
dj-database-url>=2.1.0
